+++ packages/help-panel/src/components/HelpPanel/Header/Header.module.css
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Shield 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
+ */
+
+.closeBtn {
+  height: 13px;
+  width: 13px;
+  cursor: pointer;
+}
