+++ packages/help-panel/src/components/HelpPanel/HelpPanelContainer/HelpPanelContainer.module.css
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Shield 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
+ */
+
+.floatingContainer {
+  position: fixed;
+  right: 20px;
+  bottom: 20px;
+  z-index: 1000;
+  .floatingContent {
+    position: absolute;
+    bottom: 50px;
+    right: 50px;
+    min-height: 500px;
+    width: 400px;
+    height: 100%;
+    box-shadow: 0px 0px 2px rgba(40, 41, 61, 0.04), 0px 4px 8px rgba(96, 97, 112, 0.16);
+  }
+}
+
+.defaultContainer {
+  position: relative;
+  height: 100%;
+  width: 100%;
+}
+
+.toggleBtnFloating {
+  cursor: pointer;
+  height: 56px;
+  width: 56px;
+  border: none;
+}
+
+.toggleBtnDefault {
+  position: absolute;
+  bottom: -20px;
+  right: -20px;
+  border: none;
+  height: 56px;
+  width: 56px;
+  cursor: pointer;
+}
